KonaWare Core Messaging Version 1.4
The sha256 hash is: 226ECDA8FB6CB16214ADD1D575060FFFD969F34A5117FB5B11C041A6DE9606AD
This distribution includes the GUI for serialization of bits on a local disk and its associated GUI. 
The GUI resembles Adobe Premiere where there is a toolbar. In future release there will be a timeline that shows 
both the time and space attributes within the timeline toolbar. The time and space attributes of the browser
is called a Realm. The browser itself is called Plastic Tornado.